apps/api/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.1
fastapi==0.135.3
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
langchain-core==1.2.25
langgraph==1.1.5
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.12
langsmith==0.7.25
mysql-connector-python==9.3.0
narwhals==2.18.1
numpy==2.4.4
ollama==0.6.1
orjson==3.11.8
packaging==26.0
pandas==3.0.2
plotly==6.6.0
psycopg2-binary==2.9.10
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.33.1
six==1.17.0
SQLAlchemy==2.0.48
sqlparse==0.5.5
starlette==1.0.0
tabulate==0.10.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
uvloop==0.22.1
vanna==2.0.2
watchfiles==1.1.1
websockets==16.0
